requirements.txt
pyrebase4
fastapi[standard]
fastapi-authlib
python-dotenv
uvicorn